Move port handlers into their own directories
