+++ src/js/port-handlers/firebase-handlers.js
@@ -1,9 +1,9 @@
 import firebase from 'firebase/app';
 import 'firebase/database';
 import _ from 'lodash';
-import app from './elm-init';
+import app from '../elm-init';
 
-import { config } from './firebase-config.js';
+import { config } from '../firebase-config.js';
 
 const fb = firebase.initializeApp(config()).database();
